requirements.txt
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-json-logger==2.0.2
uWSGI>=2.0.19.1
